+++ src/lib.rs
@@ -2,8 +2,10 @@ pub use tuple::*;
 pub use canvas::*;
 pub use colour::*;
 pub use matrix::*;
+pub use transformers::*;
 
 pub mod tuple;
 pub mod canvas;
 pub mod colour;
 pub mod matrix;
+pub mod transformers;